+++ service/controller/azure/resource_set.go
@@ -27,6 +27,7 @@ import (
 	"github.com/giantswarm/cluster-operator/service/controller/azure/key"
 	"github.com/giantswarm/cluster-operator/service/controller/controllercontext"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/app"
+	"github.com/giantswarm/cluster-operator/service/controller/resource/appversionlabel"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/certconfig"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/clusterconfigmap"
 	"github.com/giantswarm/cluster-operator/service/controller/resource/encryptionkey"
@@ -117,6 +118,20 @@ func newResourceSet(config resourceSetConfig) (*controller.ResourceSet, error) {
 		}
 
 		appResource, err = toCRUDResource(config.Logger, ops)
+		if err != nil {
+			return nil, microerror.Mask(err)
+		}
+	}
+
+	var appVersionLabelResource resource.Interface
+	{
+		c := appversionlabel.Config{
+			GetClusterConfigFunc: getClusterConfig,
+			G8sClient:            config.K8sClient.G8sClient(),
+			Logger:               config.Logger,
+		}
+
+		appVersionLabelResource, err = appversionlabel.New(c)
 		if err != nil {
 			return nil, microerror.Mask(err)
 		}
@@ -277,9 +292,8 @@ func newResourceSet(config resourceSetConfig) (*controller.ResourceSet, error) {
 		certConfigResource,
 		clusterConfigMapResource,
 		kubeConfigResource,
-
-		// appResource is executed after migration resources.
 		appResource,
+		appVersionLabelResource,
 	}
 
 	// Wrap resources with retry and metrics.