+++ CHANGELOG.md
@@ -7,6 +7,10 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+### Added
+
+- Add appversionlabel resource to update version labels for optional app CRs.
+
 ## [0.23.19] - 2020-12-03
 
 ### Fixed